+++ bluegroup-project-afristream.php
@@ -84,7 +84,7 @@ function afristream_portal_shortcode( $atts ) {
 	wp_add_inline_style( 'bluegroup-project-afristream', '.dashboard-right{padding:0 !important;}' );
 
 	return sprintf(
-		'<div class="afristream-portal" data-afristream-portal data-default-tab="%s" data-show-sport="%s" data-endpoint="%s" data-editor-endpoint="%s" data-detail-endpoint="%s" data-credentials-endpoint="%s" data-affiliate-endpoint="%s" data-apps-url="%s" data-rest-nonce="%s"></div>',
+		'<div class="afristream-portal" data-afristream-portal data-default-tab="%s" data-show-sport="%s" data-endpoint="%s" data-editor-endpoint="%s" data-detail-endpoint="%s" data-credentials-endpoint="%s" data-affiliate-endpoint="%s" data-apps-url="%s" data-rest-nonce="%s" data-home-url="%s"></div>',
 		esc_attr( $atts['default_tab'] ),
 		esc_attr( $atts['show_sport'] ),
 		esc_url( rest_url( 'afristream/v1/watch' ) ),
@@ -93,7 +93,8 @@ function afristream_portal_shortcode( $atts ) {
 		esc_url( rest_url( 'afristream/v1/credentials' ) ),
 		esc_url( rest_url( 'afristream/v1/affiliate' ) ),
 		esc_url( add_query_arg( 'ver', AFRISTREAM_PORTAL_VERSION, plugins_url( 'data/apps.json', __FILE__ ) ) ),
-		esc_attr( wp_create_nonce( 'wp_rest' ) )
+		esc_attr( wp_create_nonce( 'wp_rest' ) ),
+		esc_url( home_url( '/' ) )
 	);
 }
 add_shortcode( 'afristream_portal', 'afristream_portal_shortcode' );